refactor: diInjector 메서드 분리
