+++ di/src/main/java/com/woowacourse/di/DiContainer.kt
@@ -71,4 +71,38 @@ class DiContainer(private val parentContainer: DiContainer? = null) {
 
         return constructor.call(*parameters.toTypedArray())
     }
+
+    fun <T : Any> injectFields(instance: T) {
+        val clazz = instance::class
+
+        checkInjectAnnotations(clazz)
+
+        clazz.java.declaredFields.forEach { field ->
+            if (field.isAnnotationPresent(Inject::class.java)) {
+                field.isAccessible = true
+                val fieldAnnotations = field.annotations.toList()
+
+                val fieldInstance =
+                    if (fieldAnnotations.hasQualifier()) {
+                        val qualifier =
+                            fieldAnnotations.qualifierAnnotation()
+                                ?: throw IllegalArgumentException()
+                        matchByQualifier(qualifier)
+                    } else {
+                        match(field.type.kotlin)
+                    }
+                field.set(instance, fieldInstance)
+            }
+        }
+    }
+
+    fun <T : Any> checkInjectAnnotations(clazz: KClass<T>) {
+        clazz.constructors.forEach { constructor ->
+            constructor.parameters.forEach { parameter ->
+                if (!parameter.hasAnnotation<Inject>()) {
+                    throw IllegalArgumentException()
+                }
+            }
+        }
+    }
 }
